Add clever_seeker to deal with aggregator
The aggregator has an innate attribute to eaily placed
closer to controller, which will easily cause inbalance
if monitors are previously balanced without aggregator
consideration.

As a result, we adopt a "Reservoir" or "sandhill" method
of reserve rooms for aggregators.

This is still a draft, no fully tested yet. Only make sure
it is runnable.